Extract hex library and make a simple build script
