+++ set1/exercise1.rs
@@ -1,4 +1,4 @@
-use std::{num::ParseIntError};
+extern crate libpal;
 
 // Look up table that has all 64 value for encoding plus the '=' sign for the end piece 
 const LOOKUP_TABLE: [char; 64] = ['A', 'B', 'C', 'D', 'E', 'F', 'G', 'H', 'I', 'J', 'K', 'L', 'M', 'N', 'O', 'P','Q', 'R', 'S', 'T', 'U', 'V', 'W', 'X', 'Y', 'Z', 'a', 'b', 'c', 'd', 'e', 'f', 'g', 'h', 'i', 'j', 'k', 'l', 'm', 'n', 'o', 'p', 'q', 'r', 's', 't', 'u', 'v', 'w', 'x', 'y', 'z', '0', '1', '2', '3', '4', '5', '6', '7', '8', '9', '+', '/'];
@@ -40,16 +40,9 @@ fn EncodeChunk(vec: &Vec<u8>) -> Vec<char>
 	return output;
 }
 
-// Convect Hex String into Byte representation
-pub fn decode_hex(s: &str) -> Result<Vec<u8>, ParseIntError> {
-	(0..s.len())
-		.step_by(2)
-		.map(|i| u8::from_str_radix(&s[i..i + 2], 16))
-		.collect()
-}
-
 // Transforms input bytes into Base 64 representation
-pub fn hex_to_base64(s: &Vec<u8>) -> Result<Vec<char>, ParseIntError> {
+pub fn hex_to_base64(s: &Vec<u8>) -> Vec<char> 
+{
 	let mut output = Vec::new();
 	for i in (0..s.len()).step_by(3)
 	{
@@ -58,17 +51,17 @@ pub fn hex_to_base64(s: &Vec<u8>) -> Result<Vec<char>, ParseIntError> {
 		println!("{:?}, len is: {}, i {}, end {}", chunk, chunk.len(), i, end);
 		output.extend(EncodeChunk(chunk));
 	}
-	return Ok(output);
+	return output;
 }
 
 // Use Facing function
 // Takes Hex String as Input and returns Base64 string
 fn HexToBase64(input: &str) -> String
 {
-	let byte_vector = decode_hex(input);
+	let byte_vector = libpal::decode_hex(input);
 	println!("{:?}", byte_vector);
 	let output = hex_to_base64(&byte_vector.unwrap());
-	let string: String = output.unwrap().iter().cloned().collect();
+	let string: String = output.iter().cloned().collect();
 	return string;
 }
 
@@ -90,6 +83,8 @@ fn HexToBase64Test()
 	input = "49276d206b696c6c696e6720796f757220627261696e206c696b65206120706f69736f6e6f7573206d757368726f6f6d";
 	base64_result = HexToBase64(input);
 	assert_eq!(base64_result, "SSdtIGtpbGxpbmcgeW91ciBicmFpbiBsaWtlIGEgcG9pc29ub3VzIG11c2hyb29t");
+
+	println!("All tests pass!");
 }
 
 fn main()
